[refactor] 시간 함수 이동 및 수정
