+++ service/lib/fotmatted-hour-time.ts
@@ -8,4 +8,23 @@ const generateTimeOptions = (): {value: string; label: string}[] => {
   return times;
 };
 
-export default generateTimeOptions;
+const findOverlappingSchedules = (schedules: {date?: string; startTime?: string; endTime?: string}[]) => {
+  return schedules
+    .map((a, i) =>
+      schedules.some(
+        (b, j) =>
+          i !== j &&
+          a.date === b.date &&
+          a.startTime &&
+          a.endTime &&
+          b.startTime &&
+          b.endTime &&
+          ((a.startTime < b.endTime && b.startTime < a.endTime) || (a.startTime === b.startTime && a.endTime === b.endTime)),
+      )
+        ? i
+        : -1,
+    )
+    .filter(i => i !== -1);
+};
+
+export {generateTimeOptions, findOverlappingSchedules};
